Rename orangepi3b.csc to orangepi3b.eos
Ethernet is supposedly broken by design.

While Xunlong is working on a hw revision of this board it is to be expected that it will be released under the same name which will then ultimately contribute to users confusion why on some boards eth works and on others not. Therefore get rid of this one altogether.